Fix four import-time bugs from the type-hint refactor
Sorting the imports in filter/__init__.py put name_filter before
result_filter, and the filter submodules reached ResultSubFilter through
the package __init__ rather than the module defining it. Importing
mikeio1d then failed outright. They now import from .result_filter.

res1d.__all__ names and QueryData in the query converter were moved into
TYPE_CHECKING blocks, but both are needed at runtime: `import *` broke,
and convert_time_series_id_to_query raised NameError.

ResultQuantity.timeseries_id built a ValueError without raising it, so an
unattached quantity returned None instead of failing.
