+++ tests/test_result_quantity.py
@@ -1,0 +1,23 @@
+"""Tests for mikeio1d.result_network.result_quantity.ResultQuantity."""
+
+from types import SimpleNamespace
+
+import pytest
+
+from mikeio1d.result_network import ResultQuantity
+
+
+@pytest.fixture
+def quantity_outside_a_network():
+    """A ResultQuantity that was never added to a ResultNetwork.
+
+    Only data_item is read during construction, so the rest can be left unset.
+    """
+    data_item = SimpleNamespace(Quantity=SimpleNamespace(Id="WaterLevel"))
+    return ResultQuantity(result_location=None, data_item=data_item, res1d=None)
+
+
+def test_timeseries_id_raises_when_not_added_to_a_network(quantity_outside_a_network):
+    """The error used to be built and dropped, so None came back instead."""
+    with pytest.raises(ValueError, match="must be added to a ResultNetwork"):
+        quantity_outside_a_network.timeseries_id
